fix: prevent the application from crashing if a whole day occurs
If an event lasts for the whole day, it doesn't have "start-date" and
"end-date" in the JSON from khal.  Given that and that the values in the
fields are strings, the Date couldn't be parsed from those fields,
because the program accessed bytes from the empty strings.

An error is returned in case an empty strings occurs in the fields.

The errors from SDL implement std::fmt::Debug trait.  Therefore, they
can be printed.

All of the captions on the week view are grouped in the `Week`
structure.