feat(application): the event titles are rendered with some margin
In order to prevent setting the margin in two places, all of title had
to be rendered in one place.  Therefore, the titles of "All Day" events
had to be rendered together with the titles of regular events.  That
caused the refactoring of the crate calendar.  The refactoring involved
the moving of the Error to root of the crate.  Also, that caused
creating of a function which registers the titles of the events in
TextRegistry.