fix(calendar): the placing of collided events is corrected
Two events collide if one starts before the other finishes. A third
event joins the collision if it starts before any other event in the
collision finishes.

To visualize the collision, the events are placed side by side.  The
picture is similar to lanes on the road.  An event, which is about to
join the collision, finds any event which finishes before it starts.  If
such an event is found, the joining event is set below it.  Otherwise,
one more lane is added, and the joining event takes that lane.

Example:

We have two events: "One" 10:00 - 11:00 and "Two" 10:30 - 11:30.
The third event is coming up.  Let's say it's "Three" 11:00 - 12:00.

The picture is the following.  "Two" stays beside the event "One",
because "Two" starts before "One" finishes.  "Three" stays below "One",
because "Three" starts after "One" finishes.

Technically, the equation of the decision whether one event starts
before the other finishes is following:

    diff = end_time(event2) - start_time(event1)
    diff < 0, stay beside
    diff >= 0, stay below

However, the bug was in the fact that the terms were swapped by mistake.

    diff = start_time(event1) - end_time(event2)

The bug can't be seen in the example above.  Consider the following
example:

We have three events:
1. "One" 10:00 - 11:00.
2. "Two" 10:15 - 11:15.
3. "Three" 10:30 - 11:30.

The fourth event is coming up: "Four" 11:15 - 12:15.  It collides with
only with the event "Three".  Therefore, it can stay below either "One"
or "Two".

By design, it has to stay below "Two", because the gap between "One"
and "Four" is bigger.  To figure it out, "Four" has to be compare with
both of them.  The process is a loop which iterates both "One" and
"Two", calculating the differences according the equation from above.
The issue is that the initial value of the difference was the result of
subtraction of the default closest end which is Infinity.  Therefore,
the subtraction looks Infinity - start_time("Four"); as a result the
difference is Infinity.  It allows the algorithm to guarantee that the
difference calculated between events in the iterations is smaller.

With the broken equation it looks this way

start_time("Four") = 11:15
end_time("One") = 11:00
end_time("Two") = 11:15
closest_end = Infinity

First iteration:

diff = 11:00 - 11:15 => - 15 (min)
-15 < (Infinity - 11:00) => true; therefore closest_end = 11:00

Second iteration:

diff = 11:15 - 11:15 => 0 (min).
0 < (11:00 - 11:15) => false; therefore closest_end = 11:00 (BUG)

The bug caused "Four" to stay below "One".  However, only swapping
doesn't solve the problem.  Let's see the calculation one more time
with the fixed terms

First iteration:

diff = 11:15 - 11:00 => 15 (min)
15 < (11:00 - Infinity) => false; therefore closest_end = Infinity (BUG)

Second iteration:
diff = 11:15 - 11:15 => 0 (min)
0 < (11:00 - Infinity) => false; therefore closest_end = Infinity (BUG)

For that reason, the default closest end became -Infinity.