feat(application, calendar):  long events update
A long event is an event which does not finish within a day.  Namely,
its start date does match its end date.  Those events are drawn at the
top of the grid.  However, those events which shorter than a day (24
hours) are stored as two short events.  Given that, in the end of
processing the data from Khal, there are two arrays of events.  The
first one contains the short events, the second one contains long
events.

The structure `Item` is renamed to `Event`. The structure uses actual
types `Date` and `Time` instead of their strings.  Which means they are
parsed right away from input JSON.