feat(calendar, application): clashes in long events
The overlagging long events take the space at the top of the table.  The
space they are taking varies, it depends on the number of clashing
events.

As the logic affects the render, the clash detection had to be moved
from `calendar::render` to `calendar::obtain`.  The detection required
to parse the dates and times of the events into [`calendar::Date`] and
[`calendar::Time`] respectively.

The structure [`calendar::Minutes`] is introduced to address the
comparison of time ranges of the events.  That allowed event clashes to
be determined.