fix(calendar): wrong date difference calculation fixed
It was wrong, because the leap year years were calculated since 1970.
The diffirence was divided by 4.  Given that and the year 1974, we had
one extra day.  Which is fine, because 1972 was a leap year.  However,
given 1973 year, we had 0 days.  Which is incorrect, because 1972 is
still there.

Therefore, the leap years are calculated since 1972.